+++ utils/utils.go
@@ -17,9 +17,9 @@ type QueryCondition struct {
 }
 
 type XYZ struct {
-	X int `json:"x"`
-	Y int `json:"y"`
-	Z int `json:"z"`
+	X float64 `json:"x"`
+	Y float64 `json:"y"`
+	Z float64 `json:"z"`
 }
 
 type RequestParam interface {
